example/requirements.txt
Django>=1.11,<2.0
django-tenant-schemas==1.9.0
